+++ build.js
@@ -23,16 +23,14 @@ program.option("-n --no-build").parse(process.argv);
       await fs.copy(file, path.join(__dirname, config.directories.app, file.replace("src/", "")));
     }
 
-    const styles = [
-      "bulma/css/bulma.min.css",
-      "bulma-switch/dist/css/bulma-switch.min.css",
-      "jquery/dist/jquery.slim.min.js"
-    ];
-    for (let file of styles) {
-      const src = path.join(__dirname, "node_modules", file);
-      const base = path.parse(file).base
-      await fs.copy(src, path.join(__dirname, config.directories.app, base));
-    }
+    // const styles = [
+    //   "jquery-validation/dist/jquery.validate.min.js"
+    // ];
+    // for (let file of styles) {
+    //   const src = path.join(__dirname, "node_modules", file);
+    //   const base = path.parse(file).base
+    //   await fs.copy(src, path.join(__dirname, config.directories.app, base));
+    // }
 
     if (!program.build) return;