+++ src/main/log.ts
@@ -2,11 +2,12 @@ import log from "electron-log";
 import fs from "fs";
 import path from "path";
 
-function archiveLog(file: string): void {
-  var info = path.parse(file);
+function archiveLog(file: log.LogFile): void {
+  const oldPath = file.toString();
+  const info = path.parse(oldPath);
   try {
     const iso = new Date().toISOString().replace(/[-:]/g, "");
-    fs.renameSync(file, path.join(info.dir, info.name + '-' +  iso + info.ext));
+    fs.renameSync(oldPath, path.join(info.dir, info.name + '-' +  iso + info.ext));
   } catch (e) {
     // @ts-ignore
     log.transports.console({
@@ -19,6 +20,7 @@ function archiveLog(file: string): void {
 
 log.transports.console.level = 'debug';
 log.transports.file.level = 'info';
+//@ts-ignore
 log.transports.file.archiveLog = archiveLog;
 log.info(process.argv);